Fix broken 'getSelector'; add 'setSelectorBroken' to handle broken output
The old version of 'getSelector' was returning an (N x m*N) array, which
of course includes all the extra zeros needed in the 'flat' version of
the TLL network. Also added is a 'setSelectorBroken' that takes this
broken output and uses it to correctly set the selector matrix as
intended.