Retool class to store TLL using only kernel, bias and selector sets
This commit adds a special function to create a Keras modeal, and it moves a bunch of once implicitly Keras functions to variants with "Keras" in their name. Once a Keras model is created, methods for setting local linear functions/selector sets should sync automatically with the Keras model and vice versa.