Begin helper methods for optimized Keras export
The assembleAdjacency method constructs (in parallel) a list of subsets
of local linear functions formed by taking unions of the lists of
subsets provided as its arguments. In particular, though, it only
returns subsets of local linear functions that are repeated in at least
two selector sets. The resulting list can be fed back into another call
of assembleAdjacency in order to find repeated subsets of a larger size.