Implement some kind of realization procedure for min terms
This essentially uses the sets produced by the iterative matching of
earlier functions as "ingredients" to build each of the selection sets.
Priority is given to sets with a large "full" weight: i.e., the total
number of reuses of the set plus the number of reuses of its constituent
sets (but only those sets assembled by adding one element at a time
according to the iterative matching mentioned above).

This is obviously highly non-optimal, but I hope that overall procedure
yields a decent amount of min-term reuse for a relatively low
computational cost.